Doc: Replaced QSet<double> with QSet<int> in QList::fromSet() example
We don't implement qHash() for double, so let's not use QSet<double> in
the documentation.

+++ doc/src/snippets/code/src_corelib_tools_qlistdata.cpp
@@ -190,10 +190,10 @@ QVector<QString> vect = list.toVector();
 
 
 //! [23]
-QSet<double> set;
-set << 20.0 << 30.0 << 40.0 << ... << 70.0;
+QSet<int> set;
+set << 20 << 30 << 40 << ... << 70;
 
-QList<double> list = QList<double>::fromSet(set);
+QList<int> list = QList<int>::fromSet(set);
 qSort(list);
 //! [23]
